pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "veritas"
version = "0.1.0"
description = "An adaptive web research agent with step-by-step verification"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Veritas Team"},
]
keywords = [
    "research",
    "agent",
    "web",
    "verification",
    "ai",
    "llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.0.0",
    "sentence-transformers>=2.2.0",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
]
all = [
    "tiktoken>=0.5.0",
    "anthropic>=0.8.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/veritas"
Documentation = "https://github.com/yourusername/veritas#readme"
Repository = "https://github.com/yourusername/veritas"
Issues = "https://github.com/yourusername/veritas/issues"

[tool.hatch.build.targets.wheel]
packages = ["veritas"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "N", "D", "UP"]
ignore = ["D100", "D104"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
